moved common and constants
