+++ src/__legacy/common.ts
