+++ src/__legacy/iterate-repos/deploy.ts
@@ -1,10 +1,10 @@
 import { flags } from '@oclif/command'
-import { BaseCommand } from '../../common'
+import { BaseCommand } from '../common'
 import { deploy } from '../repo/deploy'
-import { getRepositories } from '../../common'
+import { getRepositories } from '../common'
 import Listr from 'listr'
 import { join } from 'path'
-import { docPathInRepo } from '../../constants'
+import { docPathInRepo } from '../../lib/constants'
 
 export default class ReposBuild extends BaseCommand {
   static description =
